Render map of all game spaces
Adds a map that displaces all of the current game spaces in the list.

DOES NOT LOOK GOOD ON MOBILE. Needs major work for use on mobile. Probably needs to just go List only or Map only.

Still need to be able to click on a map item and see the details, and click on the left list and pan the map.